* release/0.22.0: (161 commits)
  Update Changelog
  Version 0.22.0
  Add a jacobian method to atlas::Projection object
  travis: Use nproc for faster compilation
  travis: Update CMake to 3.18.4
  travis: Update OSX version
  Fix tools/install-pgi.sh download link
  Grid convenience constructors taking Projection
  Formatting
  Simplify version C++ file
  Gmsh control output of LAND/WATER elements
  Move Topology to util namespace, created BitflagsView
  atlas-grids outputs footprint of grid
  Gmsh output of idx_t coordinate fields
  Support mesh::PartitionPolygon::outputPythonScript to output polygon in "xy" or "lonlat" coordinates (default="xy")
  Tool atlas-meshgen now uses by default the Grid's recommended MeshGenerator and Partitioner
  Added constructor `Mesh::Mesh( const Grid& )`
  Added new function `MeshGenerator::generate(Mesh,Partitioner)`
  Added function `string MeshGenerator::type()`
  Polygon can return lonlat points in addition to xy points
  ...

+++ src/atlas/functionspace/Spectral.h
@@ -127,7 +127,7 @@ public:
                 }
             }
             else {
-                const int nb_zonal_wavenumbers{zonal_wavenumbers.size()};
+                const int nb_zonal_wavenumbers{static_cast<int>( zonal_wavenumbers.size() )};
                 atlas_omp_parallel_for( int jm = 0; jm < nb_zonal_wavenumbers; ++jm ) {
                     const int m = zonal_wavenumbers( jm );
                     for ( int n = m; n <= truncation; ++n ) {
@@ -153,7 +153,7 @@ public:
                 }
             }
             else {
-                const int nb_zonal_wavenumbers{zonal_wavenumbers.size()};
+                const int nb_zonal_wavenumbers{static_cast<int>( zonal_wavenumbers.size() )};
                 atlas_omp_parallel_for( int jm = 0; jm < nb_zonal_wavenumbers; ++jm ) {
                     const int m = zonal_wavenumbers( jm );
                     for ( int n = m; n <= truncation; ++n ) {
